app_service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
python-dotenv==1.0.0
httpx==0.25.0
requests==2.31.0
elasticsearch>=8.10.0
sentence-transformers>=2.2.2
chromadb>=0.4.18
langchain>=0.0.345
langchain-community>=0.0.16
numpy>=1.24.0
selenium>=4.14.0
beautifulsoup4>=4.12.2
